% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_configure_environment.R
\name{configure_environment}
\alias{configure_environment}
\title{Configure Conda Environment for barbac}
\usage{
configure_environment(
  env_name = "barbac_env",
  channels = c("conda-forge", "bioconda"),
  confirm = TRUE
)
}
\arguments{
\item{env_name}{Name of the conda environment to create. Default is "barbac_env".}

\item{channels}{List of conda channels to use (default: c("conda-forge", "bioconda"))}

\item{confirm}{If FALSE, it will skip the interactive menu. Default is TRUE.}
}
\value{
Prints progress messages and returns the environment path if successful.
}
\description{
Configure Conda Environment for barbac
}
