% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_barcodes.R
\name{extract_barcodes}
\alias{extract_barcodes}
\title{Extract Barcodes from BAM File}
\usage{
extract_barcodes(
  bam_file,
  ref_name = "Reference_barcodes",
  start_pos = 54,
  end_pos = 78
)
}
\arguments{
\item{bam_file}{Character string. Path to the BAM file from which to extract barcodes.}

\item{ref_name}{Character string. Name of the reference to be used. Default is "Reference_barcodes".}

\item{start_pos}{Numeric. Start position of the barcode in the reference. Default is 54.}

\item{end_pos}{Numeric. End position of the barcode in the reference. Default is 78.}
}
\value{
Character string. Path to the generated barcode file.
}
\description{
This function extracts barcode sequences from a BAM file based on the specified reference name and position range.
}
