% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_barcs.R
\name{group.compmat}
\alias{group.compmat}
\title{Group Comparison Function}
\usage{
group.compmat(df, method = "lv", distance)
}
\arguments{
\item{df}{A dataframe or tibble with at least columns 'barcode' and 'counts'.}

\item{method}{Character, method to be used for string distance computation.
Accepts any method provided by the \code{stringdist::stringdistmatrix} function.
Default is "lv".}

\item{distance}{Numeric distance threshold for comparison.}
}
\value{
A modified dataframe with grouped data.
}
\description{
This function compares groups in a dataframe based on a specified distance and
groups them accordingly.
}
\examples{
\dontrun{
result <- group.compmat(test, distance=2)
}

}
