% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Map2Reference.R
\name{map_and_process}
\alias{map_and_process}
\title{Map and Process Fastq File}
\usage{
map_and_process(fastq_file, reference_file, out_dir)
}
\arguments{
\item{fastq_file}{Path to the fastq file.}

\item{reference_file}{Path to the reference file.}

\item{out_dir}{Directory to save the processed files.}
}
\value{
Path to the sorted bam file.
}
\description{
This function takes a fastq file, maps it to a given reference file,
and then processes the result using samtools.
}
\examples{
\dontrun{
map_and_process("path/to/fastq", "path/to/reference", "output/directory/")
}
}
