% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_run_fastqc.R
\name{run_fastqc}
\alias{run_fastqc}
\title{Run FastQC on all FASTQ files listed in a samples.csv file}
\usage{
run_fastqc(sample_csv)
}
\arguments{
\item{sample_csv}{Path to the `samples.csv` file.}
}
\value{
A character vector of FastQC system commands executed.
}
\description{
This function reads a `samples.csv` file from a given path, extracts `R1` and optionally `R2` FASTQ paths,
and runs FastQC on all unique files. The output is saved under `results/fastQC/` in the same directory as the CSV.
}
