% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_run_pear.R
\name{run_pear_merge}
\alias{run_pear_merge}
\title{Merge paired-end reads using PEAR}
\usage{
run_pear_merge(sample_csv)
}
\arguments{
\item{sample_csv}{Full path to `samples.csv`}
}
\value{
Character vector of PEAR commands run.
}
\description{
This function reads a `samples.csv` file from the given path and merges reads using PEAR.
If only `R1` is available, the sample is skipped.
Output is saved to `results/merged/` inside the same folder.
}
