% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fastqc.R
\name{run_fastqc}
\alias{run_fastqc}
\title{Run FastQC on all fastq files in the sample table}
\usage{
run_fastqc(sample_table, output_dir = "fastQC")
}
\arguments{
\item{sample_table}{A data.frame or tibble with columns \code{R1} and \code{R2} pointing to FASTQ files}

\item{output_dir}{Directory where FastQC output will be written. Default is "fastQC"}
}
\value{
Runs FastQC via system calls. Returns a character vector of commands run.
}
\description{
Run FastQC on all fastq files in the sample table
}
