% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_multiqc.R
\name{run_multiqc}
\alias{run_multiqc}
\title{Run MultiQC to summarize FastQC reports}
\usage{
run_multiqc(input_dir = "fastQC", output_dir = "multiQC")
}
\arguments{
\item{input_dir}{The directory where FastQC output is located}

\item{output_dir}{The directory where MultiQC summary should be saved}
}
\value{
Runs MultiQC and returns the command used
}
\description{
Run MultiQC to summarize FastQC reports
}
