% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pear.R
\name{run_pear_merge}
\alias{run_pear_merge}
\title{Merge paired-end reads using PEAR}
\usage{
run_pear_merge(sample_table, output_dir = "merged")
}
\arguments{
\item{sample_table}{A data.frame or tibble with \code{sample}, \code{R1}, and \code{R2} columns}

\item{output_dir}{Where merged reads will be stored. Default is "merged"}
}
\value{
Character vector of PEAR commands run
}
\description{
Merge paired-end reads using PEAR
}
